Added an abstract savable implementation. Keywords: cscvs:MAIN-169
Author: dustin
Date: 2003/01/07 07:04:21


Added the TransactionListener which will be notified after the
transaction is commited.

Saver uses an IdentityHashMap now for storing the objects it's seen
rather than a proper Map implementation.  This will allow objects
considering themselves to be equal to exist in the Saver transaction chain
and all of the save method will be called.

imported from: dustin@spy.net--java/spyjar--sfcvs--1.0--patch-169