I was attempting to add support for decaying averages, ... Keywords: cscvs:MAIN-237
Author: dustin
Date: 2003/06/11 07:31:18

I was attempting to add support for decaying averages, but it did not
seem to be working very well when I was using the following algorithm:

a = {(a \times df) + v\over{df + 1}}

Where a is the average, df is the decay factor, and v is the new
value.  I believe this was due to floating point precision and the rates I
was working with.

I suppose I could use the RingBuffer for a similar calculation, but at
this point, I'm just going to let it get less accurate over time.

imported from: dustin@spy.net--java/spyjar--sfcvs--1.0--patch-237